+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed handling of private and protected class constants
+
 ## v0.10.21 (2023-11-09)
 * Improved deferred binding argument handling